backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.0

# MongoDB
pymongo==4.6.0
dnspython==2.6.0

# AI & Embeddings
google-generativeai==0.3.2

# Vector Database
chromadb==0.4.22

# File Processing
pypdf==4.0.1
python-docx==1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Production Dependencies
gunicorn==21.2.0  # Production WSGI server
requests>=2.31.0  # For health checks
python-json-logger==2.0.7  # Structured logging